+++ src/main/java/uniandes/edu/co/proyecto/repositorio/OrdenCompraRepository.java
@@ -1,5 +1,6 @@
 package uniandes.edu.co.proyecto.repositorio;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.Date;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -15,7 +16,7 @@ public interface OrdenCompraRepository extends JpaRepository<OrdenCompra, Intege
 
     @Query(value = "SELECT * FROM ordencompra", nativeQuery = true)
     Collection<OrdenCompra> darOrdenesCompra();
-    
+
 //inicialización de cada atributo dentro del sistema
     @Query(value = "SELECT * FROM ordencompra WHERE ORDENCOMPRA_ID = :id", nativeQuery = true)
     OrdenCompra darOrdenCompra(@Param("id") int id);
@@ -23,7 +24,7 @@ public interface OrdenCompraRepository extends JpaRepository<OrdenCompra, Intege
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO ordencompra (precio, fechaEntrega, estado, FechaCreacion, ordenCompra_id, proveedor_proveedor_id, sucursal_sucursal_id) VALUES (:precio, :fechaEntrega, 'Vigente', SYSDATE, ORDENCOMPRA_ORDENCOMPRA_ID_SEQ.nextval, :proveedor_id, :sucursal_id)", nativeQuery = true)
-    void insertarOrdenCompra(@Param("precio") Float precio, @Param("fechaEntrega") Date fechaEntrega, @Param("proveedor_id") Integer proveedor_id, @Param("sucursal_id") Integer sucursal_id);
+    void insertarOrdenCompra(@Param("precio") BigDecimal precio, @Param("fechaEntrega") Date fechaEntrega, @Param("proveedor_id") Integer proveedor_id, @Param("sucursal_id") Integer sucursal_id);
     
 
     @Modifying
